[EC-507] Move IDeleteGroupCommand to Groups folder
